feat(c-api) Move `value` into its own module.
