+++ lib/c-api/src/wasm_c_api/externals/function.rs
@@ -1,6 +1,5 @@
-use super::super::{
-    wasm_functype_t, wasm_store_t, wasm_trap_t, wasm_val_inner, wasm_val_t, wasm_valkind_enum,
-};
+use super::super::value::{wasm_val_inner, wasm_val_t};
+use super::super::{wasm_functype_t, wasm_store_t, wasm_trap_t, wasm_valkind_enum};
 use std::convert::TryInto;
 use std::ffi::c_void;
 use std::ptr::NonNull;
